Add enc28j60 ethernet driver.
There is no license included with the enc28j60 code, but the author
Iain Derrington allowed me to use it "as you wish in either a commercial or
non commercial fashion.".

The code was modified to work with the r0ket and without UIP,
some safety checks disabled per EXTRA_SAFE define to save a few bytes.