Add (broken) video streaming app, extend usb-cdc.
usb-cdc was extended to allow for a callback upon receiving data.
The current callback is too stupid, the code should rather receive data first
and then call the callback with the number of received bytes.

Currently the code does not compile, as bss is 102 bytes too large.
About 32 bytes could be saved using the improved callback described above.